+++ junit-jupiter-kotlin/src/main/resources/META-INF/services/org.junit.jupiter.engine.support.MethodAdapterFactory
@@ -1,1 +1,1 @@
-org.junit.jupiter.kotlin.KotlinSuspendFunctionAdapterFactory
+org.junit.jupiter.kotlin.coroutines.KotlinSuspendFunctionAdapterFactory
